Completed Daap Draft
Entirely functional Dapp
Requires CSS work
Requires additional tests and test editing to reflect changes to contracts

+++ Maats/contracts/StoreAuction.sol
@@ -46,7 +46,7 @@ contract StoreAuction is StorePurchasing{
   /// @dev for testing the successful creation of an auciton
 
   function getAuctionExists(uint auctionId)public constant returns(uint64){
-    uint currentTime = auctionIdToAuction[auctionId].startedAt;
+    uint64 currentTime = auctionIdToAuction[auctionId].startedAt;
     if(currentTime != 0){
       return currentTime;
     }else{
